+++ Encoding/NVEnc.vb
@@ -946,7 +946,7 @@ Public Class NVEnc
                 Case "avs"
                     sourcePath = p.Script.Path
 
-                    If includePaths AndAlso FrameServerHelp.IsAviSynthPortableUsed Then
+                    If includePaths AndAlso FrameServerHelp.IsPortable Then
                         ret += " --avsdll " + Package.AviSynth.Path.Escape
                     End If
                 Case "nvhw"