+++ Encoding/QSVEnc.vb
@@ -414,7 +414,7 @@ Public Class QSVEnc
                 Case "avs"
                     sourcePath = p.Script.Path
 
-                    If includePaths AndAlso FrameServerHelp.IsAviSynthPortableUsed Then
+                    If includePaths AndAlso FrameServerHelp.IsPortable Then
                         ret += " --avsdll " + Package.AviSynth.Path.Escape
                     End If
                 Case "qshw"